% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metapopcode.R
\name{calcarea}
\alias{calcarea}
\title{Calculates patch area}
\usage{
calcarea(habsuit, thresh = 0, asraster = TRUE)
}
\arguments{
\item{habsuit}{a SpatRaster of habitat suitability with unsuitable areas set to NA}

\item{asraster}{optional logical indicating whether to return area as a SpatRaster (TRUE)
or as a vector of areas.}
}
\value{
A SpatRaster or vector of areas (Ha) of each habitat patch
}
\description{
Calculates patch area in hectares
}
\details{
the x and y units of `habsuit` must be in metres.
}
