% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metapopcode.R
\name{calcconectivity}
\alias{calcconectivity}
\title{Calculates suitability-weighted inter-patch distances}
\usage{
calcconectivity(
  habsuit,
  thresh = 0.5,
  alpha = 10,
  aos = 0.1,
  minden = 0.01,
  maxden = 20,
  Oj = NA,
  asraster = TRUE
)
}
\arguments{
\item{habsuit}{a SpatRaster of habitat suitability with unsuitable areas set to NA}

\item{thresh}{theshold suitability value below which patches are assumed unoccupied}

\item{alpha}{inverse of average dispersal distance (km)}

\item{aos}{parameter scaling `habsuit` logistically transformed to population density (individuals per ha)}

\item{minden}{minimum possible density value (individuals per Ha)}

\item{maxden}{maximum possible density value (individuals per Ha)}

\item{Oj}{optional vector indicating which patches are occupied (1) or unoccopied (0)
asraster optional logical indicating whether to return area as a SpatRaster (TRUE)}

\item{asraster}{optional logical indicating whether to return connectivity as a SpatRaster (TRUE)
or as a vector of areas.}
}
\value{
A SpatRaster or vector of connectvity of each habitat patch to other occupied patches
}
\description{
Calculates a suitability-weighted measure of connectivity
of each habitat patch to other occupied habitat patches.
}
\details{
To calculate connectivity, first the suitability-weighted distance of each
pixel in patch i to each pixel in patch j is calculated and averaged. A matrix of
The parameter alpha is then used to define a dispersal kernel representing the likelihood of
individuals from patch j reaching each focal patch i. For a given focal patch, this value
is then summed and weighted by patch carrying capacity to give a value that scales
with the expected total number of immigrants o any given focal patch
}
