% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{.fitone}
\alias{.fitone}
\title{Fits one iteration of hybrid metapopulation-sdm model}
\usage{
.fitone(
  occdata,
  landcover,
  conpredictors,
  habsuit,
  Ar,
  Sr,
  aos = 0.1,
  thresh = 0.5,
  minden = 0.01,
  maxden = 20,
  Bayesian = FALSE
)
}
\description{
Fits one iteration of hybrid metapopulation-sdm model
}
