% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metapopcode.R
\name{fitmetapop}
\alias{fitmetapop}
\title{Parameterises a metapopulation model}
\usage{
fitmetapop(metapopdata, alphamin = 0.1, alphamax = 10, n = 100)
}
\arguments{
\item{metapopdata}{metapopdata data,frame with columns x & y (coordinates of
patches), A (area) & pa (presence absence as 1 & 0)}

\item{alphamin}{minimum possible value of alpha}

\item{alphamax}{maximum possible value of alpha}

\item{n}{number of alphas to try between alphamin & alphamax  (ignored if alphamin = alphamax)}
}
\value{
a list of the following:
\describe{
 \item{mu}{extinction probability of a patch of unit size (same units as A)}
 \item{x}{scaling of extinction risk with patch area}
 \item{ygamma}{parameter controlling how fast the colonisation probability approaches unity with increasing }
 \item{alpha}{inverse of average dispersal distance (same units as x and y)}
}
}
\description{
Derives parameters for standard IFM metapopulation model searching
for best alpha

. See vignette for further details of fitting a metapopulation model
and the meaning of parameters. If alpha is known, set `alphamin` to be the same
value as `alphamax`
}
