% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metapopcode.R
\name{fitmetapopsdm}
\alias{fitmetapopsdm}
\title{Fit hybrid metapopulation-species distribution model}
\usage{
fitmetapopsdm(
  occdata,
  landcover,
  conpredictors,
  alpha = 10,
  aos = 0.1,
  bwgt = 0,
  tol = 0.01,
  maxiter = 100,
  minden = 0.01,
  maxden = 20,
  Bayesian = FALSE
)
}
\arguments{
\item{occdata}{a data.frame of occurrence data with x & y coordinates as returned
by [subsample()].}

\item{landcover}{a SpatRaster landcover classes.}

\item{conpredictors}{a single or multi-layer SpatRaster of other continuous predictors}

\item{alpha}{inverse of average dispersal distance (km) - used to calculate connectivity}

\item{aos}{parameter scaling `habsuit` logistically transformed to population density (individuals per ha)}

\item{bwgt}{back-weighting used during iteration process in range 0 to <1. Higher = slower, but more likely to converge}

\item{tol}{tolerence limited for convergence (Maximum difference in predicted probability of occurance between sucessive iterations)}

\item{maxiter}{maximum number of iterations - will return results}

\item{minden}{minimum potential density in suitable habitat (individuals per hectare)}

\item{maxden}{maximum density (individuals per hectare)}

\item{Bayesian}{optional logical indicating whether to use a normal glm (FALSE - much faster)
or constrain the x parameter by setting Bayesian priors (TRUE - much slower)}
}
\value{
a list of two SpatRasters: (1) habsuit- the habitat suitability and (2) pocco -
estimated probability of occurence.
}
\description{
Iteratively fits a hybrid incidentfunction-species distribution model
}
