% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{habsuit}
\alias{habsuit}
\title{A 50 m resolution SpatRaster of habitat suitability for a hypothetical woodland species}
\format{
A PackedSpatRaster object with 404 rows and 476 columns
}
\source{
derived from `landcover` and `veghgt`
}
\usage{
habsuit
}
\description{
A spatial dataset of habitat suitability for the Lizard Peninsula: an area bounded by 159170,
182970, 10740, 30940 (xmin, xmax, ymin, ymax) using the Ordance Survey GB Grid Reference
system (CRS: 27700). Different land cover types are presented by numeric values as given in
the inbuilt dastaset `lclass`.
}
\keyword{datasets}
