% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lclass}
\alias{lclass}
\title{Data.frame giving legend for landcover}
\format{
An object of class \code{data.frame} with 21 rows and 3 columns.
}
\usage{
lclass
}
\description{
A data,frame with the following columns
\describe{
 \item{Habitat}{Description of habitat type}
 \item{Class}{Numeric Value corresponding to values in inbuilt dataset `landcover`}
 }
}
\keyword{datasets}
