% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{metapopdata}
\alias{metapopdata}
\title{Simulated species occurrences in habitat patches across the Lizard Peninsula}
\format{
An object of class \code{data.frame} with 506 rows and 4 columns.
}
\usage{
metapopdata
}
\description{
A data,frame with the following columns
\describe{
  \item(x){OSGB grid reference - Easting}
  \item(y){OSGB grid reference - Northing}
  \item(pa){presence (1) or absence (0)}
  \item{Area}{Patch size (ha)}
}
}
\keyword{datasets}
