% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metapopcode.R
\name{MetaPopSim}
\alias{MetaPopSim}
\title{Simulates metapopulation dynamics in a real landscape}
\usage{
MetaPopSim(
  habsuit,
  mu,
  x,
  alpha,
  ygamma,
  aos,
  timesteps = 100,
  minden = 0.01,
  maxden = 20,
  asprob = TRUE
)
}
\arguments{
\item{habsuit}{a SpatRaster of habitat suitability with unsuitable areas set to NA}

\item{mu}{extinction probability in patch of unit carrying capacity.}

\item{x}{parameter scaling patch carrying capacity to extinction}

\item{alpha}{inverse of average dispersal distance (km)}

\item{ygamma}{parameter scaling colonisation probability to connectivity}

\item{aos}{parameter scaling `habsuit` logistically transformed to population density (individuals per ha)}

\item{minden}{minimum possible density value (individuals per Ha)}

\item{maxden}{maximum possible density value (individuals per Ha)}

\item{asprob}{optional logical indicating whether to return expected probability of
occurrence (TRUE) or presence absence (FALSE)}

\item{timestep}{number of tiem-steps over which to run the model.}
}
\value{
A SpatRaster of presences / absences or occurrence probabilities.
}
\description{
Runs a stochastic patch occupancy model given a set of metapopulation
input parameters
}
