% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/popcode.R
\name{PopDynamicsSim}
\alias{PopDynamicsSim}
\title{Simulates spatially explicit population model}
\usage{
PopDynamicsSim(
  popden,
  birthrate,
  survival,
  avdispdist,
  fracdisp,
  emmsurv,
  Nt = NA
)
}
\arguments{
\item{popden}{multi-layer SpaRaster of maximum potential population densities in each timestep}

\item{birthrate}{multi-layer SpaRaster of number of births as a function of environmental
suitability within each timestep}

\item{survival}{multi-layer SpaRaster of number of survival rates as a function of environmental
suitability within each timestep}

\item{avdispdist}{average dispersal distance capability (km)}

\item{fracdisp}{fraction of population emigrating from each population annually}

\item{Nt}{optionally initial population size. Set at carrying capacity if unknown
stochastically (TRUE) or determenistically (FALSE)}

\item{em_surv}{fraction of emmigrants that typically survive}
}
\value{
a SpatRaster of predicting population density within each time-step
}
\description{
Simulates population dynamics over a landscape, including metapopulation
dynamics

. See vignette for further details
}
