% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{subsample}
\alias{subsample}
\title{Extracts sub-sample of points from simulated metapopulation model output}
\usage{
subsample(occo, landcover, n)
}
\arguments{
\item{occo}{SpatRaster of predicted probability of occupancy as returned by [MetaPopSim()]}

\item{landcover}{SpatRaster of land cover type}

\item{n}{number of points ot extract}
}
\value{
a data.frame with the following columns
\describe {
 \item{x}{randomly generated x-coordinate}
 \item{y}{randomly generated y-coordinate}
 \item{pa}{presence or absence at location given by x and y}
}
}
\description{
Extracts a sub-sample of points from simulated metapopulation model output
for subsequent testing
}
