% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{veghgt}
\alias{veghgt}
\title{A 50 m resolution SpatRaster of vegetation height}
\format{
A PackedSpatRaster object with 404 rows and 476 columns
}
\source{
derived from LiDAR data \url{http://www.tellusgb.ac.uk/}
}
\usage{
veghgt
}
\description{
A spatial dataset of vegetation height (m) for the Lizard Peninsula: an area bounded by 159170,
182970, 10740, 30940 (xmin, xmax, ymin, ymax) using the Ordance Survey GB Grid Reference
system (CRS: 27700).
}
\keyword{datasets}
